fix(core): fix probable memory leak with WeakRefs
The #processes map on node was set to remove entries when
the #finalizer registry issued the callback for a given
context. However, since the #processes map is a strong
reference, the garbage collector *cannot* initialize the
finalization process for said context.

Wrapping the context in a WeakRef before inserting it into
the #processes map should allow the finalize callback to
run eventually.