feat(transports-socket.io): array buffer support
extracts array buffers from the source objects and replaces them with a
stub. Appends the actual buffers to the end of the emit signal. On the
receiving end, once all of the buffers are available, the stubs are
replaced with references to the received buffers.

The emit signal transmits each argument separately, which means an emit
call can contain mixed argument types (objects, strings, and buffers),
and each will be transmitted in the manner most appropriate, all the
while properly reconstructing all of the terms on the receiving side.

By extracting the array buffers, we enable socket.io to perform the
transmission of large binary date much more efficiently than if it were
to be stringified and then parsed on the receiving end.