fix(serializer-json): better descent control
tried to general purpose it, but that lead to several gaps due to
inconsistencies in the way deserialize and serialize want to handle
their respective outputs.

In the serialize case, for instance, all of the child components should
be processed *after* we convert to an object structure. We attempted to
handle this by manually invoking the revive function on child elements.
This works great if the child elements are also OTP types, but if
they're native JSON types, we don't process them in the revive callback.
Hence, if a tuple contains an object that contains a list, the nested
list would not be processed.

Conversely, when deserializing, nested types should be replaced *before*
their container type, so that their true parent type can be instantiated
with the proper values in place. However, since the general case
attempted to process the *original* value, it would contain
still-serialized children. At that point, the walker would need to learn
how to walk the OTP types.

This conflict is solved most easily by breaking out the logic that
performs the transformation and customizing it to each case.

For serialize, perform the transform on the original object as before,
but then push the resulting object to the top of the stack. This
subsequently processes and replaces the tuple's children.

For deserialize, perform the transform on the *accumulated* object,
which has already transformed any nested types. This will initialize the
type with the proper elements in place.

There is room for further optimization in the descent. We could, for
instance, prevent serialize from walking children *at all* before
performing replacements.