fix(core): switch #links from Set to List
Manually scan for existing Pids in the list. This is necessary as the
Set implementation doesn't have a way to customize the comparison
function, using the SameValueZero algorithm. Since two Pid instances can
refer to the same process but will not have reference identity, that
makes SameValueZero a poor fit.